Code/requirements.txt
google-genai>=1.0.0
numpy
matplotlib
